fix(introduce convenience): Introduce convenience split methods and fix minor index handling in m...
# Project State

## Current Focus
Introduce convenience split methods and fix minor index handling in modal rendering.

## Completed
- [x] Add `Ctx::split_h` and `Ctx::split_v` to allow horizontal/vertical pane splitting with a closure on each pane.
- [x] Implement `SplitPane::from_rect` to create a split based on a `Rect` orientation automatically.
- [x] Correct modal drawing index calculations by casting to `usize` to avoid overflow warnings.
- [x] Update widget imports to include `Color` and `Styles` for consistent compositor usage.
- [x] Adjust `TabBar` imports to match the updated compositor API.
---
category: fix
scope: introduce convenience
---

+++ tools/tui/dracon-terminal-engine/src/framework/widgets/modal.rs
@@ -1,6 +1,6 @@
 use crate::framework::hitzone::HitZone;
 use crate::framework::theme::Theme;
-use crate::compositor::Plane;
+use crate::compositor::{Cell, Color, Plane, Styles};
 use ratatui::layout::Rect;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -69,7 +69,7 @@ impl<'a> Modal<'a> {
         for row in 1..self.height.saturating_sub(1) {
             let idx = (row * self.width) as usize;
             if idx < plane.cells.len() { plane.cells[idx].char = '│'; }
-            let idx = row * self.width + self.width - 1;
+            let idx = (row * self.width + self.width - 1) as usize;
             if idx < plane.cells.len() { plane.cells[idx].char = '│'; }
         }
